[refact] : Add MARK in Onboarding & Kitchen Views
