+++ RelaxOn/Views/Kitchen/CDListView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 struct CDListView: View {
+    // MARK: - State Properties
     @State var isActive: Bool = false
     @State var userRepositoriesState: [MixedSound] = userRepositories
     @State var selectedImageNames = (
@@ -19,6 +20,7 @@ struct CDListView: View {
     @State private var selectedMixedSoundIds: [Int] = []
     @State private var showingActionSheet = false
     
+    // MARK: - Life Cycles
     init(userRepositoriesState: [MixedSound] = userRepositories){
         Theme.navigationBarColors(background: UIColor(named: "RelaxBlack") ?? .black, titleColor: UIColor(named: "RelaxDimPurple") ?? .white)
         self.userRepositoriesState = userRepositoriesState
@@ -27,16 +29,21 @@ struct CDListView: View {
     var body: some View {
         
         VStack {
-            libraryHeader
+            LibraryHeader
             
             ScrollView(.vertical, showsIndicators: false) {
 
                 LazyVGrid(columns: Array(repeating: GridItem(.flexible(), alignment: .top), count: 2), spacing: 18) {
-                    plusCDImage.disabled(isEditMode)
+                    PlusCDImage
+                        .disabled(isEditMode)
 
                     ForEach(userRepositoriesState.reversed()){ mixedSound in
 
-                        CDCardView(data: mixedSound, audioVolumes: (baseVolume: mixedSound.baseSound?.audioVolume ?? 1.0, melodyVolume: mixedSound.melodySound?.audioVolume ?? 1.0, whiteNoiseVolume: mixedSound.whiteNoiseSound?.audioVolume ?? 1.0))
+                        CDCardView(data: mixedSound,
+                                   audioVolumes: (baseVolume: mixedSound.baseSound?.audioVolume ?? 1.0,
+                                                  melodyVolume: mixedSound.melodySound?.audioVolume ?? 1.0,
+                                                  whiteNoiseVolume: mixedSound.whiteNoiseSound?.audioVolume ?? 1.0)
+                        )
                             .disabled(isEditMode)
                             .overlay(alignment : .bottomTrailing) {
                                 if isEditMode {
@@ -112,19 +119,11 @@ struct CDListView: View {
             Text("These CDs will be deleted from your library")
         }
     }
-    
-    private func getEncodedData(data: [MixedSound]) -> Data? {
-        do {
-            let encoder = JSONEncoder()
-            let encodedData = try encoder.encode(data)
-            return encodedData
-        } catch {
-            print("Unable to Encode Note (\(error))")
-        }
-        return nil
-    }
-    
-    var libraryHeader: some View {
+}
+
+// MARK: - View Properties
+extension CDListView {
+    var LibraryHeader: some View {
         HStack {
             Text("CD LIBRARY")
                 .font(.title)
@@ -154,7 +153,7 @@ struct CDListView: View {
         }
     }
     
-    var plusCDImage: some View {
+    var PlusCDImage: some View {
         VStack(alignment: .leading) {
             NavigationLink(destination: StudioView(rootIsActive: self.$isActive), isActive: self.$isActive) {
                 ZStack {
