+++ RelaxOn/Views/Kitchen/CustomSegmentControlView.swift
@@ -8,13 +8,63 @@
 import SwiftUI
 
 public struct CustomSegmentControlView: View {
-
-    @Binding private var selection: Int
+    // MARK: - State Properties
     @State private var segmentSize: CGSize = .zero
     @State private var itemTitleSizes: [CGSize] = []
-
+    @Binding private var selection: Int
+    
+    // MARK: - General Properties
     private let items: [LocalizedStringKey]
+    
+    private var selectedItemWidth: CGFloat {
+        return itemTitleSizes.count > selection ? itemTitleSizes[selection].width : .zero
+    }
+    
+    private var xSpace: CGFloat {
+        guard !itemTitleSizes.isEmpty, !items.isEmpty, segmentSize.width != 0 else { return 0 }
+        let itemWidthSum: CGFloat = itemTitleSizes.map { $0.width }.reduce(0, +).rounded()
+        let space = (segmentSize.width - itemWidthSum) / CGFloat(items.count + 1)
+        return max(space, 0)
+    }
+    
+    // MARK: - Methods
+    private func segmentItemView(for index: Int) -> some View {
+        guard index < self.items.count else {
+            return EmptyView().eraseToAnyView()
+        }
 
+        let isSelected = self.selection == index
+
+        return
+            Text(items[index])
+                .font(.caption)
+                .foregroundColor(isSelected ? .white : .gray)
+                .background(BackgroundGeometryReader())
+                .onPreferenceChange(SizePreferenceKey.self) {
+                    itemTitleSizes[index] = $0
+                }
+                .onTapGesture { onItemTap(index: index) }
+                .eraseToAnyView()
+    }
+
+    private func onItemTap(index: Int) {
+        guard index < self.items.count else { return }
+        self.selection = index
+    }
+
+    private func selectedItemHorizontalOffset() -> CGFloat {
+        guard selectedItemWidth != .zero, selection != 0 else { return 0 }
+
+        let result = itemTitleSizes
+            .enumerated()
+            .filter { $0.offset < selection }
+            .map { $0.element.width }
+            .reduce(0, +)
+
+        return result + xSpace * CGFloat(selection)
+    }
+    
+    // MARK: - Life Cycles
     public init(items: [LocalizedStringKey],
                 selection: Binding<Int>) {
         self._selection = selection
@@ -53,52 +103,5 @@ public struct CustomSegmentControlView: View {
 
             }
         }
-    }
-
-    private var selectedItemWidth: CGFloat {
-        return itemTitleSizes.count > selection ? itemTitleSizes[selection].width : .zero
-    }
-
-    private func segmentItemView(for index: Int) -> some View {
-        guard index < self.items.count else {
-            return EmptyView().eraseToAnyView()
-        }
-
-        let isSelected = self.selection == index
-
-        return
-            Text(items[index])
-                .font(.caption)
-                .foregroundColor(isSelected ? .white : .gray)
-                .background(BackgroundGeometryReader())
-                .onPreferenceChange(SizePreferenceKey.self) {
-                    itemTitleSizes[index] = $0
-                }
-                .onTapGesture { onItemTap(index: index) }
-                .eraseToAnyView()
-    }
-
-    private func onItemTap(index: Int) {
-        guard index < self.items.count else { return }
-        self.selection = index
-    }
-
-    private var xSpace: CGFloat {
-        guard !itemTitleSizes.isEmpty, !items.isEmpty, segmentSize.width != 0 else { return 0 }
-        let itemWidthSum: CGFloat = itemTitleSizes.map { $0.width }.reduce(0, +).rounded()
-        let space = (segmentSize.width - itemWidthSum) / CGFloat(items.count + 1)
-        return max(space, 0)
-    }
-
-    private func selectedItemHorizontalOffset() -> CGFloat {
-        guard selectedItemWidth != .zero, selection != 0 else { return 0 }
-
-        let result = itemTitleSizes
-            .enumerated()
-            .filter { $0.offset < selection }
-            .map { $0.element.width }
-            .reduce(0, +)
-
-        return result + xSpace * CGFloat(selection)
     }
 }
