+++ RelaxOn/Views/Kitchen/VolumeSlider.swift
@@ -9,44 +9,31 @@ import SwiftUI
 import Foundation
 
 struct VolumeSlider<Component: View>: View {
-    
+    // MARK: - State Properties
     @Binding var value: Float
+    
+    // MARK: - General Properties
+    let viewBuilder: (CustomSliderComponents) -> Component
     var range: (Float, Float)
     var knobWidth: CGFloat?
-    let viewBuilder: (CustomSliderComponents) -> Component
     
-    init(value: Binding<Float>, range: (Float, Float), knobWidth: CGFloat? = nil,
-         _ viewBuilder: @escaping (CustomSliderComponents) -> Component
-    ) {
-        _value = value
-        self.range = range
-        self.viewBuilder = viewBuilder
-        self.knobWidth = knobWidth
-    }
-    
+    // MARK: - Methods
     private func onDragChange(_ drag: DragGesture.Value,_ frame: CGRect) {
         let width = (knob: Float(knobWidth ?? frame.size.height), view: Float(frame.size.width))
         let xrange = (min: Float(0), max: Float(width.view - width.knob))
         var value = Float(drag.startLocation.x + drag.translation.width) // knob center x
-        value -= 0.5*width.knob // offset from center to leading edge of knob
+        value -= 0.5 * width.knob // offset from center to leading edge of knob
         value = value > xrange.max ? xrange.max : value // limit to leading edge
         value = value < xrange.min ? xrange.min : value // limit to trailing edge
         value = value.convert(fromRange: (xrange.min, xrange.max), toRange: range)
         self.value = value
     }
     
-    
     private func getOffsetX(frame: CGRect) -> CGFloat {
         let width = (knob: knobWidth ?? frame.size.height, view: frame.size.width)
         let xrange: (Float, Float) = (0, Float(width.view - width.knob))
         let result = self.value.convert(fromRange: range, toRange: xrange)
         return CGFloat(result)
-    }
-    
-    var body: some View {
-        return GeometryReader { geometry in
-            self.view(geometry: geometry) // function below
-        }
     }
     
     private func view(geometry: GeometryProxy) -> some View {
@@ -66,6 +53,22 @@ struct VolumeSlider<Component: View>: View {
             knob: CustomSliderModifier(name: .knob, size: knobSize, offset: offsetX))
         
         return ZStack { viewBuilder(modifiers).gesture(drag) }
+    }
+    
+    // MARK: - Life Cycles
+    init(value: Binding<Float>, range: (Float, Float), knobWidth: CGFloat? = nil,
+         _ viewBuilder: @escaping (CustomSliderComponents) -> Component
+    ) {
+        _value = value
+        self.range = range
+        self.viewBuilder = viewBuilder
+        self.knobWidth = knobWidth
+    }
+    
+    var body: some View {
+        return GeometryReader { geometry in
+            self.view(geometry: geometry) // function below
+        }
     }
 }
 
@@ -88,19 +91,7 @@ struct CustomSliderModifier: ViewModifier {
     func body(content: Content) -> some View {
         content
             .frame(width: size.width)
-            .position(x: size.width*0.5, y: size.height*0.5)
+            .position(x: size.width * 0.5, y: size.height * 0.5)
             .offset(x: offset)
-    }
-}
-
-extension Float {
-    func convert(fromRange: (Float, Float), toRange: (Float, Float)) -> Float {
-        // Example: if self = 1, fromRange = (0,2), toRange = (10,12) -> solution = 11
-        var value = self
-        value -= fromRange.0
-        value /= Float(fromRange.1 - fromRange.0)
-        value *= toRange.1 - toRange.0
-        value += toRange.0
-        return value
     }
 }
